TST: Test basic functionality of sane players.
