+++ test/test_player.py
@@ -4,7 +4,7 @@ import unittest
 from pelita.datamodel import CTFUniverse, east, stop, west
 from pelita.game_master import GameMaster
 from pelita.player import *
-from pelita.players import NQRandomPlayer, RandomPlayer
+from pelita.players import RandomPlayer
 
 
 class TestAbstractPlayer:
@@ -457,38 +457,6 @@ class TestRandomPlayerSeeds:
         random_numbers_c = [player.rnd.randint(0, 10000) for player in players_c]
 
         assert random_numbers_a != random_numbers_c
-
-
-class TestNQRandom_Player:
-    def test_demo_players(self):
-        test_layout = (
-        """ ############
-            #0#.   .# 1#
-            ############ """)
-        team = [
-            SimpleTeam(NQRandomPlayer()),
-            SimpleTeam(NQRandomPlayer())
-        ]
-        gm = GameMaster(test_layout, team, 2, 1)
-        gm.play()
-        assert gm.universe.bots[0].current_pos == (1, 1)
-        assert gm.universe.bots[1].current_pos == (9, 1)
-
-    def test_path(self):
-        test_layout = (
-        """ ############
-            #  . # .# ##
-            # ## #  # ##
-            #0#.   .##1#
-            ############ """)
-        team = [
-            SimpleTeam(NQRandomPlayer()),
-            SimpleTeam(NQRandomPlayer())
-        ]
-        gm = GameMaster(test_layout, team, 2, 7)
-        gm.play()
-        assert gm.universe.bots[0].current_pos == (4, 3)
-        assert gm.universe.bots[1].current_pos == (10, 3)
 
 
 class TestSpeakingPlayer:
